requirements.txt
#alembic
asyncpg
bcrypt==4.1.2
cloudinary
fastapi
fastapi_jwt_auth
fastapi_mail
google-cloud-storage
jinja2
pandas
passlib[bcrypt]
pathlib
pillow
psycopg2-binary
pydantic
pydantic-settings
python-dateutil
python-dotenv
python-multipart
python-jose[cryptography]
qrcode
starlette
sqlalchemy
uuid
uvicorn
cloud-sql-python-connector==1.2.4
functions-framework==3.3.0
pg8000==1.29.8
